added tostify

+++ src/pages/home/Home.js
@@ -6,9 +6,12 @@ import { useSelector } from "react-redux";
 import { useDispatch } from "react-redux";
 import { deleteUrl, findOne, getData } from "../../redux/actions";
 import { useHistory } from "react-router-dom";
+import { ToastContainer, toast } from "react-toastify";
 
 const Home = () => {
-  const data = useSelector((state) => state.data);
+  const data = useSelector((state) => state?.data);
+  const auth = localStorage.getItem("auth");
+  console.log(auth);
   const dispatch = useDispatch();
   const history = useHistory();
   useEffect(() => {
@@ -16,11 +19,20 @@ const Home = () => {
     console.log("data", data);
   }, []);
   const handleEdit = (id) => {
-    dispatch(findOne(id, history));
+    if (auth === "true") {
+      dispatch(findOne(id, history));
+    } else {
+      toast.error("You can't access");
+    }
   };
   const handleDelete = (id) => {
-    dispatch(getData());
-    dispatch(deleteUrl(id));
+    if (auth === "true") {
+      dispatch(deleteUrl(id));
+      dispatch(getData());
+      toast.success("Deleted Successfully");
+    } else {
+      toast.error("You can't access");
+    }
   };
   return (
     <>
@@ -29,8 +41,9 @@ const Home = () => {
         style={{ minHeight: "calc(100vh - 104px)" }}
       >
         <div className="container mx-auto py-3 px-2 flex flex-col gap-y-4">
-          {data &&
-            data?.data?.map((item) => (
+          {data?.data &&
+            data?.data.length > 0 &&
+            data?.data.map((item) => (
               <div className="bg-white p-3 shadow rounded-lg" key={item._id}>
                 <div className="res-flex-col flex items-center justify-between">
                   <div>
@@ -61,55 +74,9 @@ const Home = () => {
                 </div>
               </div>
             ))}
-
-          {/* <div className="bg-white p-3 shadow rounded-lg ">
-            <div className="res-flex-col flex items-center justify-between">
-              <div>
-                <p className="font-semibold text-lg">Title here</p>
-                <div className=" w-full border rounded py-2 px-5 bg-blue-100 text-gray-400">
-                  https://pink-grasshopper-c7ecs1q8.ws-us25.gitpod.io/
-                </div>
-              </div>
-              <div className="flex items-center justify-start gap-x-2">
-                <NavLink to="/" target="_blank">
-                  <button className="outline-0 bg-blue-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                    Open
-                  </button>
-                </NavLink>
-                <button className="outline-0 text-2xl bg-green-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                  <FaEdit />
-                </button>
-                <button className="outline-0 text-2xl bg-red-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                  <MdDelete />
-                </button>
-              </div>
-            </div>
-          </div>
-          <div className="bg-white p-3 shadow rounded-lg ">
-            <div className="res-flex-col flex items-center justify-between">
-              <div>
-                <p className="font-semibold text-lg">Title here</p>
-                <div className=" w-full border rounded py-2 px-5 bg-blue-100 text-gray-400">
-                  https://pink-grasshopper-c7ecs1q8.ws-us25.gitpod.io/
-                </div>
-              </div>
-              <div className="flex items-center justify-start gap-x-2">
-                <NavLink to="/" target="_blank">
-                  <button className="outline-0 bg-blue-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                    Open
-                  </button>
-                </NavLink>
-                <button className="outline-0 text-2xl bg-green-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                  <FaEdit />
-                </button>
-                <button className="outline-0 text-2xl bg-red-600 text-white  py-2 px-5 rounded shadow ml-auto hover:shadow-xl transition-all">
-                  <MdDelete />
-                </button>
-              </div>
-            </div>
-          </div> */}
         </div>
       </div>
+      <ToastContainer />
     </>
   );
 };
